+++ tests/test_streamer.py
@@ -69,7 +69,7 @@ def test_drop():
         assert list(s) == [0, 1, 2, 4, 6, 7]
 
     with Streamer(data) as s:
-        assert list(s.drop_first_n(6)) == [6, 7]
+        assert list(s.drop_if(lambda i, x: i < 6)) == [6, 7]
 
     with Streamer((2, 3, 1, 5, 4, 7)) as s:
         assert list(s.drop_if(lambda i, x: x > i)) == [1, 4]
@@ -89,11 +89,11 @@ def test_keep():
 
     with Streamer(data) as s:
         s.head(4)
-        s.drop_first_n(3)
+        s.drop_if(lambda i, x: i < 3)
         assert list(s) == [3]
 
     with Streamer(data) as s:
-        assert list(s.drop_first_n(3).head(1)) == [3]
+        assert list(s.drop_if(lambda i, x: i < 3).head(1)) == [3]
 
 
 def test_peek():
@@ -105,7 +105,11 @@ def test_peek():
         n = s.peek_every_nth(3).drain()
         assert n == 10
 
-    with Streamer(data).peek_random(0.5, lambda i, x: print(f'--{i}--  {x}')) as s:
+    def foo(i, x):
+        if random.random() < 0.5:
+            print(f'--{i}-- {x}')
+
+    with Streamer(data).peek(foo) as s:
         n = s.drain()
         assert n == 10