+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+Removed
+-------
+
+- `mpservice.streamer.Streamer.{drop_first_n, peek_random}`.
+
+
 Changed
 -------